+++ group_router.go
@@ -720,11 +720,7 @@ func (r *GroupRoute) Call(ctx *Context) {
 		ctx.response.Content = result[FirstOutParamOffset].Interface()
 	} else {
 		err := last.Interface().(error)
-		if ctx.response.StatusCode == 0 {
-			// 允许上层自定义错误码
-			ctx.response.StatusCode = DefaultErrorStatusCode
-		}
-		ctx.response.Content = err.Error()
+		ctx.response.StatusCode, ctx.response.Content = routeErrorHandle(ctx, err)
 	}
 }